Code Clean (#583)
* updatre format

* remove dead code

* move struct into `account_basic`

* close #576

* remove dead code

* fix test
