+++ bin/node/runtime/pangolin/src/pallets/mod.rs
@@ -1,0 +1,116 @@
+pub mod system;
+pub use system::*;
+
+pub mod babe;
+pub use babe::*;
+
+pub mod timestamp;
+pub use timestamp::*;
+
+pub mod balances;
+pub use balances::*;
+
+pub mod transaction_payment;
+pub use transaction_payment::*;
+
+pub mod authorship;
+pub use authorship::*;
+
+pub mod election_provider_multi_phase;
+pub use election_provider_multi_phase::*;
+
+pub mod staking;
+pub use staking::*;
+
+pub mod offences;
+pub use offences::*;
+
+pub mod session_historical;
+pub use session_historical::*;
+
+pub mod session;
+pub use session::*;
+
+pub mod grandpa;
+pub use grandpa::*;
+
+pub mod im_online;
+pub use im_online::*;
+
+pub mod authority_discovery;
+pub use authority_discovery::*;
+
+pub mod header_mmr;
+pub use header_mmr::*;
+
+pub mod democracy;
+pub use democracy::*;
+
+pub mod collective;
+pub use collective::*;
+
+pub mod elections_phragmen;
+pub use elections_phragmen::*;
+
+pub mod membership;
+pub use membership::*;
+
+pub mod treasury;
+pub use treasury::*;
+
+pub mod sudo;
+pub use sudo::*;
+
+pub mod claims;
+pub use claims::*;
+
+pub mod vesting;
+pub use vesting::*;
+
+pub mod utility;
+pub use utility::*;
+
+pub mod identity;
+pub use identity::*;
+
+pub mod society;
+pub use society::*;
+
+pub mod recovery;
+pub use recovery::*;
+
+pub mod scheduler;
+pub use scheduler::*;
+
+pub mod proxy;
+pub use proxy::*;
+
+pub mod multisig;
+pub use multisig::*;
+
+pub mod crab_issuing;
+pub use crab_issuing::*;
+
+pub mod crab_backing;
+pub use crab_backing::*;
+
+pub mod ethereum_relay;
+pub use ethereum_relay::*;
+
+pub mod ethereum_backing;
+pub use ethereum_backing::*;
+
+pub mod relayer_game;
+pub use relayer_game::*;
+
+pub mod relay_authorities;
+pub use relay_authorities::*;
+
+pub mod tron_backing;
+pub use tron_backing::*;
+
+pub mod evm;
+pub use evm::*;
+
+pub mod dvm;
+pub use dvm::*;
