+++ bin/node/runtime/pangolin/src/pallets/offences.rs
@@ -7,8 +7,8 @@ use sp_runtime::Perbill;
 use crate::*;
 
 frame_support::parameter_types! {
-	pub OffencesWeightSoftLimit: Weight = Perbill::from_percent(60) *
-		RuntimeBlockWeights::get().max_block;
+	pub OffencesWeightSoftLimit: Weight = Perbill::from_percent(60)
+		* RuntimeBlockWeights::get().max_block;
 }
 impl Config for Runtime {
 	type Event = Event;
