+++ bin/node/runtime/pangolin/src/pallets/scheduler.rs
@@ -7,8 +7,8 @@ use sp_runtime::Perbill;
 use crate::*;
 
 frame_support::parameter_types! {
-	pub MaximumSchedulerWeight: Weight = Perbill::from_percent(80) *
-		RuntimeBlockWeights::get().max_block;
+	pub MaximumSchedulerWeight: Weight = Perbill::from_percent(80)
+		* RuntimeBlockWeights::get().max_block;
 	pub const MaxScheduledPerBlock: u32 = 50;
 }
 impl Config for Runtime {
