+++ frame/dvm/src/account_basic.rs
@@ -1,10 +1,13 @@
-use crate::{Config, KtonBalance, KtonRemainBalance, RingBalance, RingRemainBalance};
+use crate::{Config, KtonBalance, RemainingKtonBalance, RemainingRingBalance, RingBalance};
 use darwinia_evm::{Account as EVMAccount, AccountBasic, AddressMapping};
 use evm::ExitError;
 use frame_support::ensure;
-use frame_support::traits::Currency;
+use frame_support::{storage::StorageMap, traits::Currency};
 use sp_core::{H160, U256};
-use sp_runtime::{traits::UniqueSaturatedInto, SaturatedConversion};
+use sp_runtime::{
+	traits::{Saturating, UniqueSaturatedInto},
+	SaturatedConversion,
+};
 
 pub trait RemainBalanceOp<T: Config, B> {
 	/// Get the remaining balance
@@ -19,8 +22,67 @@ pub trait RemainBalanceOp<T: Config, B> {
 	fn dec_remaining_balance(account_id: &T::AccountId, value: B);
 }
 
-pub struct DvmAccountBasic<T, C, RB>(sp_std::marker::PhantomData<(T, C, RB)>);
+pub struct RingRemainBalance;
+impl<T: Config> RemainBalanceOp<T, RingBalance<T>> for RingRemainBalance {
+	/// Get the remaining balance
+	fn remaining_balance(account_id: &T::AccountId) -> RingBalance<T> {
+		<RemainingRingBalance<T>>::get(account_id)
+	}
+	/// Set the remaining balance
+	fn set_remaining_balance(account_id: &T::AccountId, value: RingBalance<T>) {
+		<RemainingRingBalance<T>>::insert(account_id, value)
+	}
+	/// Remove the remaining balance
+	fn remove_remaining_balance(account_id: &T::AccountId) {
+		<RemainingRingBalance<T>>::remove(account_id)
+	}
+	/// Inc remaining balance
+	fn inc_remaining_balance(account_id: &T::AccountId, value: RingBalance<T>) {
+		let remain_balance =
+			<Self as RemainBalanceOp<T, RingBalance<T>>>::remaining_balance(account_id);
+		let updated_balance = remain_balance.saturating_add(value);
+		<RemainingRingBalance<T>>::insert(account_id, updated_balance);
+	}
+	/// Dec remaining balance
+	fn dec_remaining_balance(account_id: &T::AccountId, value: RingBalance<T>) {
+		let remain_balance =
+			<Self as RemainBalanceOp<T, RingBalance<T>>>::remaining_balance(account_id);
+		let updated_balance = remain_balance.saturating_sub(value);
+		<RemainingRingBalance<T>>::insert(account_id, updated_balance);
+	}
+}
 
+pub struct KtonRemainBalance;
+impl<T: Config> RemainBalanceOp<T, KtonBalance<T>> for KtonRemainBalance {
+	/// Get the remaining balance
+	fn remaining_balance(account_id: &T::AccountId) -> KtonBalance<T> {
+		<RemainingKtonBalance<T>>::get(account_id)
+	}
+	/// Set the remaining balance
+	fn set_remaining_balance(account_id: &T::AccountId, value: KtonBalance<T>) {
+		<RemainingKtonBalance<T>>::insert(account_id, value)
+	}
+	/// Remove the remaining balance
+	fn remove_remaining_balance(account_id: &T::AccountId) {
+		<RemainingKtonBalance<T>>::remove(account_id)
+	}
+	/// Inc remaining balance
+	fn inc_remaining_balance(account_id: &T::AccountId, value: KtonBalance<T>) {
+		let remain_balance =
+			<Self as RemainBalanceOp<T, KtonBalance<T>>>::remaining_balance(account_id);
+		let updated_balance = remain_balance.saturating_add(value);
+		<RemainingKtonBalance<T>>::insert(account_id, updated_balance);
+	}
+	/// Dec remaining balance
+	fn dec_remaining_balance(account_id: &T::AccountId, value: KtonBalance<T>) {
+		let remain_balance =
+			<Self as RemainBalanceOp<T, KtonBalance<T>>>::remaining_balance(account_id);
+		let updated_balance = remain_balance.saturating_sub(value);
+		<RemainingKtonBalance<T>>::insert(account_id, updated_balance);
+	}
+}
+
+pub struct DvmAccountBasic<T, C, RB>(sp_std::marker::PhantomData<(T, C, RB)>);
 impl<T: Config, C, RB> AccountBasic for DvmAccountBasic<T, C, RB>
 where
 	RB: RemainBalanceOp<T, C::Balance>,
